Add custom tooltips for annotated observations.
This change introduces an AnnotatedObservation. You can make an
annotated observation with a blank annotation, in which the standard
tooltip shows up on hover, or you can add a list of annotations to show.
In particular this allows for annotating medication montoring with
timing information.

In particular this change enables tooltips for medication monitoring.

Change-Id: Id596ac595d455996f4e9cb9b0a3e39c22b55624f
GitOrigin-RevId: fc0739108eb3d322ffb0923fa411a94ff56092c7